requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic
python-dotenv
databases[postgresql] 
asyncpg
psycopg2-binary
bcrypt==4.0.1
passlib[bcrypt]
pydantic[email]
python-jose
python-multipart
passlib